requirements.txt.txt
pandas==2.0.0
numpy==1.24.0
scikit-learn==1.3.0
fastapi==0.104.0
uvicorn==0.24.0
sqlalchemy==2.0.0
psycopg2-binary==2.9.9
streamlit==1.28.0
apache-airflow==2.7.0
great-expectations==0.18.0
python-multipart==0.0.6
requests==2.31.0
pymsteams==0.2.2
joblib==1.3.2